+++ agents/craftassist/craftassist_agent.py
@@ -44,6 +44,7 @@ from droidlet.perception.craftassist.voxel_models.subcomponent_classifier import
 )
 from droidlet.lowlevel.minecraft import craftassist_specs
 from droidlet.lowlevel.minecraft.craftassist_cuberite_utils.block_data import COLOR_BID_MAP
+from droidlet.lowlevel.minecraft import shape_helpers as sh
 
 from droidlet.event import sio
 
@@ -85,9 +86,15 @@ class CraftAssistAgent(LocoMCAgent):
         self.init_inventory()
         self.init_event_handlers()
 
+        shape_helper_dict = {
+            "shape_names": sh.SHAPE_NAMES,
+            "shape_helper": sh.SHAPE_HELPERS,
+            "bid": sh.bid(),
+            "shape_fns": sh.SHAPE_FNS
+        }
         # list of (prob, default function) pairs
         self.visible_defaults = [
-            (0.001, default_behaviors.build_random_shape),
+            (0.001, (default_behaviors.build_random_shape, shape_helper_dict)),
             (0.005, default_behaviors.come_to_player),
         ]
         self.perceive_on_chat = True
