+++ agents/craftassist/tests/test_default_behaviors.py
@@ -2,7 +2,7 @@ import unittest
 import numpy as np
 from agents.craftassist.tests.base_craftassist_test_case import BaseCraftassistTestCase
 from droidlet.interpreter.craftassist.default_behaviors import build_random_shape, come_to_player
-
+from droidlet.lowlevel.minecraft import shape_helpers as sh
 
 class TestDefaultBehavior(BaseCraftassistTestCase):
     def setUp(self):
@@ -12,9 +12,15 @@ class TestDefaultBehavior(BaseCraftassistTestCase):
         schematic = [1] * 1000
         # only build small things, otherwise test takes a long time and is likely to
         # spill out of world space (and so take forever)
+        shape_helper_dict = {
+            "shape_names": sh.SHAPE_NAMES,
+            "shape_helper": sh.SHAPE_HELPERS,
+            "bid": sh.bid(),
+            "shape_fns": sh.SHAPE_FNS
+        }
         while len(schematic) > 200:
             self.agent.memory.task_stack_clear()
-            schematic = build_random_shape(self.agent, rand_range=(1, 0, 1))
+            schematic = build_random_shape(self.agent, shape_helper_dict, rand_range=(1, 0, 1))
         # Assert that some non-zero size schematic was built
         self.assertTrue(len(schematic) > 0)
         changes = self.flush(10000)
