+++ agents/loco_mc_agent.py
@@ -369,7 +369,14 @@ class LocoMCAgent(BaseAgent):
         fn = np.random.choice(fns, p=p)
         if fn != noop:
             logging.debug("Default behavior: {}".format(fn))
-        fn(self)
+
+        if type(fn) == tuple:
+            # this function has arguments
+            f, args = fn
+            f(self, args)
+        else:
+            # run defualt
+            fn(self)
 
     def maybe_dump_memory_to_dashboard(self):
         if time.time() - self.dashboard_memory_dump_time > MEMORY_DUMP_KEYFRAME_TIME:
