+++ droidlet/interpreter/craftassist/put_memory_handler.py
@@ -6,7 +6,6 @@ import logging
 from typing import Dict, Tuple, Any, Optional
 
 from droidlet.dialog.dialogue_objects import DialogueObject, Say
-
 from droidlet.interpreter import (
     FilterInterpreter,
     ReferenceObjectInterpreter,
